+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.2] - To be released
 
 ### Fixed
-- League lookup should handle `null` response if ID was not found
+- League serializer should handle `null` response if ID was not found
+- Event serializer should handle `null` response if ID was not found
 
 ## [1.0.1] - 2021-05-04